fix(fs:remame): add error if proper file exists

+++ src/fs/rename.js
@@ -7,6 +7,9 @@ const WRONG_FILE_PATH = path.join(FILES_DIR, 'wrongFilename.txt');
 const PROPER_FILE_PATH = path.join(FILES_DIR, 'properFilename.md');
 
 const rename = async () => {
+  if (await exists(PROPER_FILE_PATH)) {
+    throw new FsError();
+  }
   if (await exists(WRONG_FILE_PATH)) {
     await fs.rename(WRONG_FILE_PATH, PROPER_FILE_PATH);
   } else {
